refactor: move compMatch binding to standalone file
